+++ core/src/test/scala/unit/kafka/server/CreateTopicsRequestTest.scala
@@ -162,7 +162,7 @@ class CreateTopicsRequestTest extends AbstractCreateTopicsRequestTest {
   def testCreateTopicsRequestVersions(quorum: String): Unit = {
     // Note: we don't run this test when in KRaft mode, because kraft does not yet support returning topic
     // configs from CreateTopics.
-    for (version <- ApiKeys.CREATE_TOPICS.oldestVersion to ApiKeys.CREATE_TOPICS.latestVersion) {
+    for (version <- ApiKeys.CREATE_TOPICS.oldestVersion.toInt to ApiKeys.CREATE_TOPICS.latestVersion.toInt) {
       val topic = s"topic_$version"
       val data = new CreateTopicsRequestData()
       data.setTimeoutMs(10000)
